Update README.md
This repository contains a comprehensive analysis of the significant data breach incident that occurred in 2023, involving the unauthorized access and exposure of personal data belonging to over 50 million citizens of Bangladesh from a government website. This project aims to dissect the incident, understand the vulnerabilities exploited, evaluate the government's response, and propose measures to bolster cybersecurity practices.